feat(globals): add p_table
this is a "special" table which if the key is not found it searches
through each key for a matching pattern. This is particularly useful for
trying to index values for something like a plugin's filetype where
there can be dozens of filetypes but a common prefix e.g.
Neogit