feat(winbar): don't show full file path
When using relative paths for winbar, the full path is used when
opening a file from a directory not relative to the file. This usually
occurs when using aliases to quickly open file. This leads to info such
as username appearing in winbar i.e. 'home/awesomeuser/somefile'. This
checks whenever this scenario happens and pops of the first two
components i.e. 'home/awesomeuser', of the file path. Hence, the
relative path starts from the home directory. Eg. .config/nvim/somefile